fix S3Client
